requirements.txt
python-libmaas
PyYAML
